Added workaround
